chore(libflux): move docs logic into its own Rust module
Prior to this change the code for generating docs from the stdlib was in
a mix of the bootstrap and flux modules. Now there is a doc module
explicitly for logic related to parsing docs from source.
