+++ libflux/go/libflux/buildinfo.gen.go
@@ -39,9 +39,10 @@ var sourceHashes = map[string]string{
 	"libflux/flux-core/src/scanner/token.rs":                                                      "5090c2ac4b341566a85f7d9ed9b48746dd2084ec2f3effdb4a7dc16cf34d44b9",
 	"libflux/flux-core/src/scanner/unicode.rl":                                                    "f923f3b385ddfa65c74427b11971785fc25ea806ca03d547045de808e16ef9a1",
 	"libflux/flux-core/src/scanner/unicode.rl.COPYING":                                            "6cf2d5d26d52772ded8a5f0813f49f83dfa76006c5f398713be3854fe7bc4c7e",
-	"libflux/flux-core/src/semantic/bootstrap.rs":                                                 "4745ffc0b6f7b7a482c3b82c74eb1582ab70b5f434e6356a96d94a5b79fab1aa",
+	"libflux/flux-core/src/semantic/bootstrap.rs":                                                 "43dafed018a404e8e5efc0e7af4d534a6c20cb4f32c3bb7fe0b3bc60e2676fca",
 	"libflux/flux-core/src/semantic/check.rs":                                                     "476db74a4479baff7091bb85a5ac2b8e9ffc07883bec92f9543a1024366ff9db",
 	"libflux/flux-core/src/semantic/convert.rs":                                                   "fc78500a57dbfb9521303a5c5140939efe583ce855cd232afd5e6d43a8817893",
+	"libflux/flux-core/src/semantic/doc.rs":                                                       "494895503b8d411be3c4ca088ce71c043f5ac3f8300f239bac79c253b215c30c",
 	"libflux/flux-core/src/semantic/env.rs":                                                       "7bbc29da93f07f7ad795e90d3b444791a195d98f6688a7bad5feffff9784902e",
 	"libflux/flux-core/src/semantic/flatbuffers/mod.rs":                                           "ed53628573204f3849f9dfa08b4aaf7b6ee0bcbb06d13109e559d385cc38ba7f",
 	"libflux/flux-core/src/semantic/flatbuffers/semantic_generated.rs":                            "c6355368fd2a16da7f907fd6607658fc3c4c722c990a225d1316673e1d80e0da",
@@ -50,7 +51,7 @@ var sourceHashes = map[string]string{
 	"libflux/flux-core/src/semantic/fresh.rs":                                                     "8c558023e20a705f2e0782969be4008ef841f21b17e72400235722d7dcde8084",
 	"libflux/flux-core/src/semantic/import.rs":                                                    "e291110f42b35a09973c5820d143231b1eeda7800a7e91638bfdc489096c79b9",
 	"libflux/flux-core/src/semantic/infer.rs":                                                     "76e0340c8ab89f45aa385febecd06ee599cdc020aca53cab5b434640018417c4",
-	"libflux/flux-core/src/semantic/mod.rs":                                                       "e2d3dfe9f80d542e4caffae54eb9f1992a21413dcede951dc28381dbf4627437",
+	"libflux/flux-core/src/semantic/mod.rs":                                                       "3cb4b9cbd51cf450e671338f7cb8114e701ff3f7f817f83b3cc80f1ebca85e55",
 	"libflux/flux-core/src/semantic/nodes.rs":                                                     "e19b389ffbb04915294c7faf2958afd7a71cde98a462a83d8f03139e160983a0",
 	"libflux/flux-core/src/semantic/sub.rs":                                                       "7a6300e22d1e3ff8f4263c17b9af6df806d7d214128d62a099d9b1ba209f5621",
 	"libflux/flux-core/src/semantic/tests.rs":                                                     "9c66626d17b86d5444afa745c213263f3d61663fcea7f4ca065573c0f477311a",
@@ -68,7 +69,7 @@ var sourceHashes = map[string]string{
 	"libflux/flux/benches/formatter.rs":                                                           "e4838c1428d45205292a9dda1759f804779a4e142db4360ec2a6be1c9ca741b8",
 	"libflux/flux/build.rs":                                                                       "06924d7c98c84ea5e38d446716c7e12cbb60981f21b1eba1e97f278e366a1173",
 	"libflux/flux/src/bin/flux-dump-docs.rs":                                                      "73c64a6cd9a4dd280fed402e5bdae52e22b0625d500943e1666f3ba7e3c36cde",
-	"libflux/flux/src/lib.rs":                                                                     "250a3fcafe945dfd07e3aea6d80a53a6ef638a87b067183afdbd7a53e885e186",
+	"libflux/flux/src/lib.rs":                                                                     "ac2364f5f219338b5fd00331f302100cdb3c19f48eca68e6664cd0fc10282cfb",
 	"libflux/flux/templates/base.html":                                                            "a818747b9621828bb96b94291c60922db54052bbe35d5e354f8e589d2a4ebd02",
 	"libflux/flux/templates/home.html":                                                            "f9927514dd42ca7271b4817ad1ca33ec79c03a77a783581b4dcafabd246ebf3f",
 	"libflux/flux/templates/package.html":                                                         "635e1d638ad1a430f3894ff0a458572ca3c4dc6b9cec2c57fe771443849e1402",
